requirements.txt
chainer==3.1.0
numpy==1.11.0
h5py==2.6.0
cupy==2.1.0
docopt==0.6.2
scipy==1.0.0
PyYAML==3.12
